added

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,17 +2,61 @@
 
 namespace App\Http\Controllers\Admin;
 use App\Models\Post;
+use App\Models\User;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Pagination\Paginator;
+
 
 class PostController extends Controller
+
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index(){
 
+        
+        
+       
+        
         $posts= Post::all();
+        $posts=Post::paginate(5);
 
-        return view('admin.post.index', compact('posts'));
+        return view('admin.post.index', compact('posts', 'posts'));
     }
+
+
+
+    public function approved($id){
+         $data=post::find($id);
+         $data->status='Approved';
+         $data->save();
+         if(auth()->user()->role_as == '1'){
+            return redirect()->back();
+         }else{
+           
+           return redirect()->back()->with('error', 'You are not admin');
+         }
+          
+
+    }
+    public function pending($id){
+        $data=post::find($id);
+        $data->status='Pending';
+        $data->save();
+
+        if(auth()->user()->role_as == '1'){
+            return redirect()->back();
+         }else{
+           
+           return redirect()->back()->with('error', 'You are not admin');
+         }
+
+   }
 
     public function create(){
         return view('admin.post.create');
@@ -22,13 +66,14 @@ class PostController extends Controller
         $validate= $request->validate([
           'title'=>['required','string'],
           'articles'=>['required','string'],
-          'status'=>['string'],
+          
 
         ]);
         $post= new Post;
         $post->title=$validate['title'];
         $post->articles=$validate['articles'];
-        $post->status=$request->status == true ? '1':'0';
+        $post->status='IN PROGRESS';
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('admin/posts')->with('message', 'Post added successfully');
@@ -37,6 +82,13 @@ class PostController extends Controller
 
     public function edit($id){
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->role_as == '1'){
+            return view('admin.post.edit',compact('post'));
+        }
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('admin/posts')->with('error', 'Unauthorized page');
+        }
         return view('admin.post.edit',compact('post'));
     }
 
@@ -59,6 +111,9 @@ class PostController extends Controller
 
     public function destroy($id){
         $post = Post::find($id);
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('admin/posts')->with('error', 'Unauthorized page');
+        }
         $post->delete();
         return redirect('admin/posts')->with('message', 'Post deleted successfully');
